+++ libmtx/mtxfile/comments.h
@@ -142,7 +142,7 @@ int mtxfile_fread_comments(
     struct mtxfile_comments * comments,
     FILE * f,
     int * lines_read,
-    int * bytes_read,
+    int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
 
@@ -159,7 +159,7 @@ int mtxfile_gzread_comments(
     struct mtxfile_comments * comments,
     gzFile f,
     int * lines_read,
-    int * bytes_read,
+    int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
 #endif
